feat: Support safepoint poll (#46)
## Related issue(s):

issue #7

## What this PR does / why we need it:

The safepoint poll is defined as JavaOp, and is inlined in the later
stage of optimization.

I use the thread local poll word to determine whether to enter a
safepoint, and leverage a runtime stub to transfer execution to the
C++-implemented safepoint handler.

Currently, only safepoint polls for back edges.

---------

Signed-off-by: taofengliu <liutaofeng.ltf@antgroup.com>